feat(api): Add skeleton of Thermocycler API class
Closes #2992
