+++ api/tests/opentrons/hardware_control/modules/test_hc_thermocycler.py
@@ -1,0 +1,6 @@
+from opentrons.hardware_control import modules
+
+
+def test_sim_initialization():
+    temp = modules.build('', 'thermocycler', True)
+    assert isinstance(temp, modules.AbstractModule)
