+++ shared-data/robot-data/moduleSpecs.json
@@ -34,5 +34,23 @@
         "displayName": "Temperature Module",
         "loadName": "tempdeck",
         "quirks": []
+    },
+    "thermocycler": {
+        "labwareOffset": {
+            "x": -0.5,
+            "y": -8.81,
+            "z": 103.44
+        },
+        "dimensions": {
+            "bareOverallHeight": 108.5,
+            "overLabwareHeight": 0
+        },
+        "calibrationPoint": {
+            "x": 0.0,
+            "y": 0.0
+        },
+        "displayName": "Thermocycler Module",
+        "loadName": "thermocycler",
+        "quirks": []
     }
 }
